requirements.txt
# Django core
Django>=5.2,<7.0
django-unfold>=0.40.0
python-dotenv==1.0.0

# Template engine (Jinja2 for Django)
Jinja2>=3.1.2

# Password hashing (for migrated Flask-Bcrypt passwords)
bcrypt>=4.0.0

# Image & file handling
Pillow>=9.5.0

# QR codes
qrcode>=7.4.2
pypng>=0.20220715.0

# PDF generation
weasyprint>=59.0

# i18n
Babel>=2.12.1

# Date/time utilities
python-dateutil>=2.8.2
pytz>=2023.3
tzlocal>=5.0.1
